Unify status chip logic
