+++ src/orders/components/OrderDetailsPage/OrderDetailsPage.tsx
@@ -8,6 +8,7 @@ import Form from "@saleor/components/Form";
 import Grid from "@saleor/components/Grid";
 import Metadata, { MetadataFormData } from "@saleor/components/Metadata";
 import PageHeader from "@saleor/components/PageHeader";
+import PageTitleWithStatusChip from "@saleor/components/PageTitleWithStatusChip";
 import Savebar from "@saleor/components/Savebar";
 import Skeleton from "@saleor/components/Skeleton";
 import { SubmitPromise } from "@saleor/hooks/useForm";
@@ -35,8 +36,8 @@ import OrderFulfilledProductsCard from "../OrderFulfilledProductsCard";
 import OrderHistory, { FormData as HistoryFormData } from "../OrderHistory";
 import OrderInvoiceList from "../OrderInvoiceList";
 import OrderPayment from "../OrderPayment/OrderPayment";
+import OrderStatusChip from "../OrderStatusChip/OrderStatusChip";
 import OrderUnfulfilledProductsCard from "../OrderUnfulfilledProductsCard";
-import Title from "./Title";
 import { filteredConditionalItems, hasAnyItemsReplaceable } from "./utils";
 
 const useStyles = makeStyles(
@@ -223,7 +224,11 @@ const OrderDetailsPage: React.FC<OrderDetailsPageProps> = props => {
             <PageHeader
               className={classes.header}
               inline
-              title={<Title order={order} />}
+              title={
+                <PageTitleWithStatusChip title={order?.number}>
+                  <OrderStatusChip order={order} />
+                </PageTitleWithStatusChip>
+              }
             >
               <CardMenu menuItems={selectCardMenuItems} />
             </PageHeader>
